+++ src/client/src/components/form/formSelect.tsx
@@ -1,15 +1,48 @@
-import { MenuItem } from "@mui/material";
+import { MenuItem, SxProps } from "@mui/material";
 import { useTranslation } from "react-i18next";
 import { Controller, useFormContext } from "react-hook-form";
-import { FormField, getFormFieldBackgroundColor } from "./form";
+import { getFormFieldError } from "./form";
+import { FormField } from "./formField";
+import { FC } from "react";
 
-export const FormSelect = props => {
-  const { fieldName, label, required, disabled, selected, values, sx, onUpdate } = props;
+export interface FormSelectProps {
+  fieldName: string;
+  label: string;
+  required?: boolean;
+  disabled?: boolean;
+  selected?: number[];
+  values?: FormSelectValue[];
+  sx?: SxProps;
+  onUpdate?: (value: number) => void;
+}
+
+export interface FormSelectValue {
+  key: number;
+  name: string;
+}
+
+export interface FormSelectMenuItem {
+  key: number;
+  value?: number;
+  label: string;
+  italic?: boolean;
+}
+
+export const FormSelect: FC<FormSelectProps> = ({
+  fieldName,
+  label,
+  required,
+  disabled,
+  selected,
+  values,
+  sx,
+  onUpdate,
+}) => {
   const { t } = useTranslation();
   const { control } = useFormContext();
 
-  var menuItems = [];
-  menuItems.push({ key: "0", value: "", label: t("reset"), italic: true });
+  const menuItems: FormSelectMenuItem[] = [];
+  menuItems.push({ key: 0, value: undefined, label: t("reset"), italic: true });
 
   if (values) {
     values.forEach(value => {
@@ -38,13 +71,9 @@ export const FormSelect = props => {
         <FormField
           select
           required={required ?? false}
-          sx={{
-            backgroundColor: getFormFieldBackgroundColor(fieldName, formState?.errors),
-            ...sx,
-          }}
-          size="small"
+          error={getFormFieldError(fieldName, formState.errors)}
+          sx={{ ...sx }}
           label={t(label)}
-          variant="outlined"
           name={field.name}
           onChange={field.onChange}
           onBlur={field.onBlur}