+++ src/client/src/pages/detail/form/borehole/geometryImport.jsx
@@ -20,7 +20,6 @@ import {
   useBoreholeGeometryMutations,
 } from "../../../../api/fetchApiV2.js";
 import { Controller, FormProvider, useForm, useWatch } from "react-hook-form";
-import { FormSelect } from "../../../../components/form/form.js";
 import CopyIcon from "../../../../assets/icons/copy.svg?react";
 import { AlertContext } from "../../../../components/alert/alertContext.tsx";
 import { StackHalfWidth } from "../../../../components/styledComponents.js";